Add progress so far
So far, I have started a class for ellipse calculations. It can do a straightforward projection onto the xy plane, and it can rotate the plane we are projecting onto in 2 axes. It's far from finished and I suspect I'll end up revamping it entirely several times, but it's a start :)